% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bind_data}
\alias{bind_data}
\title{Bind data to project}
\usage{
bind_data(project, df, ID_col = 1, pop_col = NULL, ploidy_col = NULL,
  data_cols = NULL, ploidy = 1, missing_data = -9, name = NULL,
  check_delete_output = TRUE)
}
\arguments{
\item{project}{TODO}

\item{df}{TODO}

\item{ID_col}{TODO}

\item{pop_col}{TODO}

\item{ploidy_col}{TODO}

\item{data_cols}{TODO}

\item{ploidy}{TODO}

\item{missing_data}{TODO}

\item{name}{TODO}

\item{check_delete_output}{TODO}
}
\description{
Bind data to project
}
\details{
TODO
}
\examples{
# TODO
}
