% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{change_set}
\alias{change_set}
\title{Change active parameter set}
\usage{
change_set(project, set)
}
\arguments{
\item{project}{an rmaverick project, as produced by the function 
\code{mavproject()}}

\item{set}{which set to make the new active set}
}
\description{
Change the active parameter set within an rmaverick project.
}
