% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_rmaverick_loaded}
\alias{check_rmaverick_loaded}
\title{Check that rmaverick package has loaded successfully}
\usage{
check_rmaverick_loaded()
}
\description{
Simple function to check that rmaverick package has loaded
  successfully. Prints "rmaverick loaded successfully!" if so.
}
