% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{delete_set}
\alias{delete_set}
\title{Delete parameter set}
\usage{
delete_set(project, set = NULL, check_delete_output = TRUE)
}
\arguments{
\item{project}{an rmaverick project, as produced by the function
\code{mavproject()}}

\item{set}{TODO}

\item{check_delete_output}{TODO}
}
\description{
Delete parameter set
}
\details{
TODO
}
\examples{
# TODO
}
