% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_scaffolds}
\alias{generate_scaffolds}
\title{Generate scaffolds}
\usage{
generate_scaffolds(project, iterations = NULL, n = 10, coupling_on = TRUE,
  splitmerge_on = TRUE, cluster = NULL)
}
\arguments{
\item{project}{TODO}

\item{iterations}{TODO}

\item{n}{TODO}

\item{coupling_on}{TODO}

\item{splitmerge_on}{TODO}

\item{cluster}{TODO}
}
\description{
TODO
}
\examples{
# TODO
}
