% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_ESS}
\alias{get_ESS}
\title{Get effective sample size (ESS) over all K}
\usage{
get_ESS(proj, K = NULL)
}
\arguments{
\item{proj}{TODO}

\item{K}{TODO}
}
\description{
Get effective sample size (ESS) over all K
}
\details{
TODO
}
\examples{
# TODO
}
