% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_qmatrix}
\alias{get_qmatrix}
\title{Extract q-matrix for a given analysis}
\usage{
get_qmatrix(proj, K, s = NULL)
}
\arguments{
\item{proj}{an rmaverick project, as produced by the function 
\code{mavproject()}}

\item{K}{which value of K to extract}

\item{s}{which set to extract from. Defaults to the current active set}
}
\description{
Simple function for extracting the q-matrix output from a given
  parameter set (defaults to the active set) and value of K.
}
