% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.maverick_GTI_logevidence}
\alias{is.maverick_GTI_logevidence}
\title{Determine if object is of class maverick_GTI_logevidence}
\usage{
is.maverick_GTI_logevidence(x)
}
\arguments{
\item{x}{any R object}
}
\description{
Determine if object is of class maverick_GTI_logevidence
}
