% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.maverick_GTI_path}
\alias{is.maverick_GTI_path}
\title{Determine if object is of class maverick_GTI_path}
\usage{
is.maverick_GTI_path(x)
}
\arguments{
\item{x}{any R object}
}
\description{
Determine if object is of class maverick_GTI_path
}
