% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.maverick_GTI_posterior_model}
\alias{is.maverick_GTI_posterior_model}
\title{Determine if object is of class maverick_GTI_posterior_model}
\usage{
is.maverick_GTI_posterior_model(x)
}
\arguments{
\item{x}{any R object}
}
\description{
Determine if object is of class maverick_GTI_posterior_model
}
\details{
TODO
}
\examples{
# TODO
}
