% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.maverick_loglike_quantiles}
\alias{is.maverick_loglike_quantiles}
\title{Determine if object is of class maverick_loglike_quantiles}
\usage{
is.maverick_loglike_quantiles(x)
}
\arguments{
\item{x}{any R object}
}
\description{
Determine if object is of class maverick_loglike_quantiles
}
\details{
TODO
}
\examples{
# TODO
}
