% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.maverick_qmatrix_ind}
\alias{is.maverick_qmatrix_ind}
\title{Determine if object is of class maverick_qmatrix_ind}
\usage{
is.maverick_qmatrix_ind(x)
}
\arguments{
\item{x}{any R object}
}
\description{
Determine if object is of class maverick_qmatrix_ind
}
