% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mavproject.R
\name{is.mavproject}
\alias{is.mavproject}
\title{Determine if object is of class mavproject}
\usage{
is.mavproject(x)
}
\arguments{
\item{x}{object of class \code{mavproject}}
}
\description{
Determine if object is of class mavproject
}
\details{
TODO
}
\examples{
# TODO
}
