% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{maverick_file}
\alias{maverick_file}
\title{Import file}
\usage{
maverick_file(name)
}
\arguments{
\item{name}{name of file}
}
\description{
Import file from the inst/extdata folder of this package
}
\details{
TODO
}
\examples{
# TODO
}
