% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{new_set}
\alias{new_set}
\title{Create new parameter set}
\usage{
new_set(project, name = "(no name)", admix_on = FALSE, K = 1:3,
  rungs = 11)
}
\arguments{
\item{project}{TODO}

\item{name}{TODO}

\item{admix_on}{TODO}

\item{K}{TODO}

\item{rungs}{TODO}
}
\description{
Create new parameter set
}
\details{
TODO
}
\examples{
# TODO
}
