% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.maverick_GTI_path}
\alias{plot.maverick_GTI_path}
\title{Default plot for class maverick_GTI_path}
\usage{
\method{plot}{maverick_GTI_path}(x, y, ...)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{...}{TODO}
}
\description{
TODO
}
\details{
TODO
}
\examples{
# TODO
}
