% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.maverick_GTI_posterior}
\alias{plot.maverick_GTI_posterior}
\title{Default plot for class maverick_GTI_posterior}
\usage{
\method{plot}{maverick_GTI_posterior}(x, y, ...)
}
\arguments{
\item{x}{an object of class \code{maverick_GTI_posterior}}

\item{y}{(ignored)}

\item{...}{other arguments (ignored)}
}
\description{
Default plot for class maverick_GTI_posterior
}
