% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_acf}
\alias{plot_acf}
\title{Produce MCMC autocorrelation plot}
\usage{
plot_acf(proj, K = NULL, rung = NULL, param = "alpha", col = "black")
}
\arguments{
\item{proj}{TODO}

\item{K}{TODO}

\item{rung}{TODO}

\item{param}{TODO}

\item{col}{TODO}
}
\description{
Produce MCMC autocorrelation plot of alpha or log-likelihood
}
\details{
TODO
}
\examples{
# TODO
}
