% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_alpha}
\alias{plot_alpha}
\title{Produce diagnostic plots of parameter alpha}
\usage{
plot_alpha(proj, K = NULL, col = "black")
}
\arguments{
\item{proj}{TODO}

\item{K}{TODO}

\item{col}{TODO}
}
\description{
Produce diagnostic plots of parameter alpha
}
\details{
TODO
}
\examples{
# TODO
}
