% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_density}
\alias{plot_density}
\title{Produce MCMC density plot}
\usage{
plot_density(proj, K = NULL, rung = NULL, param = "alpha",
  col = "black")
}
\arguments{
\item{proj}{TODO}

\item{K}{TODO}

\item{rung}{TODO}

\item{param}{TODO}

\item{col}{TODO}
}
\description{
Produce MCMC density plot of alpha or log-likelihood
}
\details{
TODO
}
\examples{
# TODO
}
