% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_GTI_path}
\alias{plot_GTI_path}
\title{Plot GTI path of current active set}
\usage{
plot_GTI_path(proj, K = NULL, axis_type = 1)
}
\arguments{
\item{proj}{an rmaverick project, as produced by the function 
\code{mavproject()}}

\item{K}{which value of K to produce the plot for}

\item{axis_type}{how to format the x-axis. 1 = integer rungs, 2 = values of
beta}
}
\description{
Plot GTI path of current active set
}
