% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_logevidence_K}
\alias{plot_logevidence_K}
\title{Plot log-evidence estimates over K}
\usage{
plot_logevidence_K(proj, ...)
}
\arguments{
\item{proj}{TODO}

\item{...}{TODO}
}
\description{
Plot log-evidence estimates over K
}
\details{
TODO
}
\examples{
# TODO
}
