% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_logevidence_model}
\alias{plot_logevidence_model}
\title{Plot log-evidence estimates over parameter sets}
\usage{
plot_logevidence_model(proj)
}
\arguments{
\item{proj}{an rmaverick project, as produced by the function 
\code{mavproject()}}
}
\description{
Plot log-evidence estimates over parameter sets
}
