% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_loglike_quantiles}
\alias{plot_loglike_quantiles}
\title{Plot loglike quantiles of current active set}
\usage{
plot_loglike_quantiles(proj, K = NULL, axis_type = 3,
  connect_points = FALSE, ...)
}
\arguments{
\item{proj}{TODO}

\item{K}{TODO}

\item{axis_type}{TODO}

\item{connect_points}{TODO}

\item{...}{TODO}
}
\description{
Plot loglike quantiles of current active set
}
\details{
TODO
}
\examples{
# TODO
}
