% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mc_acceptance}
\alias{plot_mc_acceptance}
\title{Plot Metropolis coupling acceptance rates}
\usage{
plot_mc_acceptance(proj, K = NULL, x_axis_type = 1)
}
\arguments{
\item{proj}{an rmaverick project, as produced by the function 
\code{mavproject()}}

\item{K}{which value of K to produce the plot for}

\item{x_axis_type}{how to format the x-axis. 1 = integer rungs, 2 = values of
the thermodynamic power.}
}
\description{
Plot Metropolis coupling acceptance rates between all adjacent
  rungs. It is important that acceptance rates are greater than zero between
  all pairs of rungs in order to give confidence that the MCMC has thoroughly
  explored the full posterior distribution. If there are any points in this
  plot where acceptance rates drop close to zero then consider changing the
  number of rungs or their distribution (see \code{rungs} and \code{GTI_pow}
  arguments in \code{run_mcmc()}, respectively).
}
