% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_posterior_model}
\alias{plot_posterior_model}
\title{Plot posterior model}
\usage{
plot_posterior_model(proj)
}
\arguments{
\item{proj}{an rmaverick project, as produced by the function 
\code{mavproject()}}
}
\description{
Plot posterior model
}
