% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_qmatrix_ind}
\alias{plot_qmatrix_ind}
\title{Plot individual-level Q-matrix of current active set}
\usage{
plot_qmatrix_ind(proj, K = NULL, ...)
}
\arguments{
\item{proj}{TODO}

\item{K}{TODO}

\item{...}{TODO}
}
\description{
Plot individual-level Q-matrix of current active set
}
\details{
TODO
}
\examples{
# TODO
}
