% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print.maverick_GTI_posterior}
\alias{print.maverick_GTI_posterior}
\title{Overload print() function for class maverick_GTI_posterior}
\usage{
\method{print}{maverick_GTI_posterior}(x, ...)
}
\arguments{
\item{x}{an object of class \code{maverick_GTI_posterior}}

\item{...}{other arguments passed to \code{print()} function}
}
\description{
Overload print() function for class maverick_GTI_posterior
}
