% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print.maverick_loglike_quantiles}
\alias{print.maverick_loglike_quantiles}
\title{Overload print() function for class maverick_loglike_quantiles}
\usage{
\method{print}{maverick_loglike_quantiles}(x, ...)
}
\arguments{
\item{x}{an object of class \code{maverick_loglike_quantiles}}

\item{...}{other arguments passed to \code{print()} function}
}
\description{
Overload print() function for class maverick_loglike_quantiles
}
\details{
TODO
}
\examples{
# TODO
}
