% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print.maverick_qmatrix_ind}
\alias{print.maverick_qmatrix_ind}
\title{Overload print() function for class maverick_qmatrix_ind}
\usage{
\method{print}{maverick_qmatrix_ind}(x, ...)
}
\arguments{
\item{x}{an object of class \code{maverick_qmatrix_ind}}

\item{...}{other arguments passed to \code{print()} function}
}
\description{
Overload print() function for class maverick_qmatrix_ind
}
\details{
TODO
}
\examples{
# TODO
}
