% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mavproject.R
\name{print.mavproject}
\alias{print.mavproject}
\title{Print function for class mavproject}
\usage{
\method{print}{mavproject}(x, ...)
}
\arguments{
\item{x}{object of class \code{mavproject}}

\item{...}{other parameters to pass to print function}
}
\description{
Overload print function for class mavproject
}
\details{
TODO
}
\examples{
# TODO
}
