% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{recalculate_evidence}
\alias{recalculate_evidence}
\title{Recalculate evidence and posterior estimates}
\usage{
recalculate_evidence(proj)
}
\arguments{
\item{proj}{an rmaverick project, as produced by the function 
\code{mavproject()}}
}
\description{
When a new value of K is added in to the analysis it affects all downstream evidence estimates that depend on this K - for example the overall model evidence integrated over K. This function therefore looks through all values of K in the active set and recalculates all downstream elements as needed.
}
