% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run main MCMC}
\usage{
run_mcmc(project, K = 3, burnin = 100, samples = 1000, rungs = 10,
  GTI_pow = 3, auto_converge = TRUE, solve_label_switching_on = TRUE,
  coupling_on = TRUE, scaffold_on = TRUE, splitmerge_on = TRUE,
  cluster = NULL, silent = !is.null(cluster))
}
\arguments{
\item{project}{an rmaverick project, as produced by the function
\code{mavproject()}}

\item{K}{the values of K to run the MCMC on}

\item{burnin}{the number of burn-in iterations}

\item{samples}{the number of sampling iterations}

\item{rungs}{the number of temperature rungs}

\item{GTI_pow}{the power used in generalised thermodynamic integration}

\item{auto_converge}{whether convergence should be assessed automatically, in which case burn-in iterations are terminated as soon as convergence is reached. Otherwise the full \code{burnin} iterations are used}

\item{solve_label_switching_on}{whether to implement the Stevens' solution to the label-switching problem}

\item{coupling_on}{whether to implement Metropolis-coupling over temperature rungs}

\item{scaffold_on}{whether to use scaffolds to improve mixing}

\item{splitmerge_on}{whether to implement a split-merge proposal}

\item{cluster}{pass in a cluster environment}

\item{silent}{whether to suppress all console output}
}
\description{
Run main MCMC
}
\details{
TODO
}
\examples{
# TODO
}
