% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mavproject.R
\name{summary.mavproject}
\alias{summary.mavproject}
\title{Print summary for class mavproject}
\usage{
\method{summary}{mavproject}(object, ...)
}
\arguments{
\item{object}{object of class \code{mavproject}}

\item{...}{other parameters to pass to summary function}
}
\description{
Overload summary function for class mavproject
}
\details{
TODO
}
\examples{
# TODO
}
