% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{example_mcmc}
\alias{example_mcmc}
\title{Example MCMC}
\usage{
example_mcmc(x, K = 3, mu_prior_mean = 0, mu_prior_var = 1000,
  sigma = 1, burnin = 100, samples = 1000, rungs = 10,
  coupling_on = TRUE, parallel_on = TRUE, num_cores = NULL)
}
\arguments{
\item{x}{TODO}

\item{K}{TODO}

\item{mu_prior_mean}{TODO}

\item{mu_prior_var}{TODO}

\item{sigma}{TODO}

\item{burnin}{TODO}

\item{samples}{TODO}

\item{rungs}{TODO}

\item{coupling_on}{TODO}

\item{parallel_on}{TODO}

\item{num_cores}{TODO}
}
\description{
Demonstrates the use of Rcpp and RcppParallel through a simple MCMC.
}
\examples{
#TODO
}
