% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{example_mcmc}
\alias{example_mcmc}
\title{Example MCMC}
\usage{
example_mcmc(x, mu_prior_mean = 0, mu_prior_var = 1000, burnin = 100,
  samples = 1000, rungs = 10, num_threads = NULL)
}
\arguments{
\item{x}{TODO}

\item{mu_prior_mean}{TODO}

\item{mu_prior_var}{TODO}

\item{burnin}{TODO}

\item{samples}{TODO}

\item{rungs}{TODO}

\item{num_threads}{TODO}
}
\description{
Demonstrates the use of Rcpp and RcppParallel through a simple MCMC.
}
\examples{
example_mcmc(1:5)
}
