% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_scaffolds}
\alias{generate_scaffolds}
\title{Generate scaffolds}
\usage{
generate_scaffolds(project, n = 10, coupling_on = TRUE,
  splitmerge_on = TRUE, cluster = NULL, silent = !is.null(cluster))
}
\arguments{
\item{project}{an rmaverick project, as produced by the function
\code{mavproject()}}

\item{n}{TODO}

\item{coupling_on}{TODO}

\item{splitmerge_on}{TODO}

\item{cluster}{TODO}

\item{silent}{TODO}
}
\description{
Generate scaffolds
}
\details{
TODO
}
\examples{
# TODO
}
