% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{new_set}
\alias{new_set}
\title{Create new parameter set}
\usage{
new_set(project, name = "(no name)", lambda = 1, admix_on = FALSE,
  alpha = 0.1, estimate_alpha = TRUE)
}
\arguments{
\item{project}{an rmaverick project, as produced by the function
\code{mavproject()}}

\item{name}{TODO}

\item{lambda}{TODO}

\item{admix_on}{TODO}

\item{alpha}{TODO}

\item{estimate_alpha}{TODO}
}
\description{
Create new parameter set
}
\details{
TODO
}
\examples{
# TODO
}
