% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.maverick_GTI_path}
\alias{plot.maverick_GTI_path}
\title{Default plot for class maverick_GTI_path}
\usage{
\method{plot}{maverick_GTI_path}(x, y, ...)
}
\arguments{
\item{x}{an object of class \code{maverick_GTI_path}}

\item{y}{how to format the x-axis. 1 = integer rungs, 2 = values of
beta}

\item{...}{other arguments (ignored)}
}
\description{
Default plot for class maverick_GTI_path
}
