% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.maverick_qmatrix_ind}
\alias{plot.maverick_qmatrix_ind}
\title{Default plot for class maverick_qmatrix_ind}
\usage{
\method{plot}{maverick_qmatrix_ind}(x, y, ...)
}
\arguments{
\item{x}{an object of class \code{maverick_qmatrix_ind}}

\item{y}{(ignored)}

\item{...}{other arguments (ignored)}
}
\description{
Default plot for class maverick_qmatrix_ind
}
