% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_alpha}
\alias{plot_alpha}
\title{Produce diagnostic plots of parameter alpha}
\usage{
plot_alpha(proj, K = NULL, col = "black")
}
\arguments{
\item{proj}{an rmaverick project, as produced by the function 
\code{mavproject()}}

\item{K}{which value of K to produce the plot for}

\item{col}{colour of the trace}
}
\description{
Produce diagnostic plots of parameter alpha
}
