% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_density}
\alias{plot_density}
\title{Produce MCMC density plot}
\usage{
plot_density(proj, K = NULL, rung = NULL, param = "alpha",
  col = "black")
}
\arguments{
\item{proj}{an rmaverick project, as produced by the function 
\code{mavproject()}}

\item{K}{which value of K to produce the plot for}

\item{rung}{which value of K to produce the plot for. Defaults to cold rung}

\item{param}{whether to produce trace plot of admixture parameter
(\code{"alpha"}) or the log-likelihood (\code{"loglike"})}

\item{col}{colour of the trace}
}
\description{
Produce MCMC density plot of alpha or log-likelihood
}
