% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_GTI_path}
\alias{plot_GTI_path}
\title{Plot GTI path of current active set}
\usage{
plot_GTI_path(proj, K = NULL, axis_type = 2, connect_points = TRUE, ...)
}
\arguments{
\item{proj}{TODO}

\item{K}{TODO}

\item{axis_type}{TODO}

\item{connect_points}{TODO}

\item{...}{TODO}
}
\description{
Plot GTI path of current active set
}
\details{
TODO
}
\examples{
# TODO
}
