% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_qmatrix}
\alias{plot_qmatrix}
\title{Plot Q-matrix of current active set}
\usage{
plot_qmatrix(proj, K = NULL, divide_ind_on = FALSE)
}
\arguments{
\item{proj}{an rmaverick project, as produced by the function 
\code{mavproject()}}

\item{K}{which value of K to produce the plot for}

\item{divide_ind_on}{whether to add dividing lines between bars}
}
\description{
Plot Q-matrix of current active set
}
