% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_qmatrix}
\alias{plot_qmatrix}
\title{Plot Q-matrix of current active set}
\usage{
plot_qmatrix(proj, K = NULL)
}
\arguments{
\item{proj}{TODO}

\item{K}{TODO}
}
\description{
Plot Q-matrix of current active set
}
\details{
TODO
}
\examples{
# TODO
}
