% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mavproject.R
\name{print.mavproject}
\alias{print.mavproject}
\title{Custom print function for class mavproject}
\usage{
\method{print}{mavproject}(x, ...)
}
\arguments{
\item{x}{object of class \code{mavproject}}

\item{...}{other arguments (ignored)}
}
\description{
Custom print function for class mavproject, printing a summary of the key elements (also equivalent to \code{summary(x)}). To do an ordinary \code{print()} of all elements of the project, use the \code{print_full()} function.
}
\examples{
# TODO
}
