% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{recalculate_evidence}
\alias{recalculate_evidence}
\title{Recalculate evidence and posterior estimates}
\usage{
recalculate_evidence(project)
}
\arguments{
\item{project}{TODO}
}
\description{
Recalculate evidence and posterior estimates of active set
}
\details{
TODO
}
\examples{
# TODO
}
