% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run main MCMC}
\usage{
run_mcmc(project, K = 3, burnin = 100, samples = 1000, rungs = 10,
  GTI_pow = 2, auto_converge = TRUE,
  converge_test = ceiling(burnin/10), solve_label_switching_on = TRUE,
  coupling_on = TRUE, cluster = NULL, pb_markdown = FALSE,
  silent = FALSE)
}
\arguments{
\item{project}{an rmaverick project, as produced by the function 
\code{mavproject()}}

\item{K}{the values of K that the MCMC will explore}

\item{burnin}{the number of burn-in iterations}

\item{samples}{the number of sampling iterations}

\item{rungs}{the number of temperature rungs}

\item{GTI_pow}{the power used in the generalised thermodynamic integration 
method. Must be greater than 1.1}

\item{auto_converge}{whether convergence should be assessed automatically 
every \code{converge_test} iterations, leading to termination of the 
burn-in phase. If \code{FALSE} then the full \code{burnin} iterations are 
used}

\item{converge_test}{test for convergence every \code{convergence_test} 
iterations if \code{auto_converge} is being used}

\item{solve_label_switching_on}{whether to implement the Stevens' solution to
the label-switching problem. If turned off then Q-matrix output will no 
longer be correct, although evidence estimates will be unaffected.}

\item{coupling_on}{whether to implement Metropolis-coupling over temperature 
rungs}

\item{cluster}{option to pass in a cluster environment (see package 
"parallel")}

\item{pb_markdown}{whether to run progress bars in markdown mode, in which 
case they are updated once at the end to avoid large amounts of output.}

\item{silent}{whether to suppress all console output}
}
\description{
Run the main rmaverick MCMC. Model parameters are taken from the
  current active parameter set, and MCMC parameters are passed in as
  arguments. All output is stored within the project.
}
